--- lang/aladino/interpreter_test.go
@@ -1,38 +1,0 @@
-// Copyright 2022 Explore.dev Unipessoal Lda. All Rights Reserved.
-// Use of this source code is governed by a license that can be
-// found in the LICENSE file.
-
-package aladino_test
-
-import (
-	"fmt"
-	"log"
-	"testing"
-
-	"github.com/reviewpad/reviewpad/v2/lang/aladino"
-	mocks_aladino "github.com/reviewpad/reviewpad/v2/mocks/aladino"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestProcessLabel(t *testing.T) {
-	mockedEnv, err := mocks_aladino.MockDefaultEnv(nil, nil)
-	if err != nil {
-		log.Fatalf("mockDefaultEnv failed: %v", err)
-	}
-
-	mockedInterpreter := &aladino.Interpreter{
-		Env: mockedEnv,
-	}
-
-	labelID := "label_id"
-	labelName := "label_name"
-	err = mockedInterpreter.ProcessLabel(labelID, labelName)
-
-	internalLabelID := fmt.Sprintf("@label:%v", labelID)
-	gotVal := mockedEnv.GetRegisterMap()[internalLabelID]
-
-	wantVal := aladino.BuildStringValue(labelName)
-
-	assert.Nil(t, err)
-	assert.Equal(t, wantVal, gotVal)
-}
